requirements.txt
colorama~=0.4.6
pytz~=2024.2
pysftp~=0.2.9
requests~=2.32.3
PyYAML~=6.0.2
psutil~=6.0.0
setuptools~=75.1.0
pexpect~=4.9.0
GitPython~=3.1.43
PyGithub~=2.4.0
python-gitlab~=4.12.2
cryptography~=43.0.1
python-dotenv~=1.0.1
nikassets~=0.7
niklibrary~=0.19